OE-19 Separate layout and JS for frontend and admin.
The CSS is still the same, could split that too if we though style.css was unnecessarily big (I think it will grow to contain a bunch of admin stuff, so I'm leaning toward yes even if it's not that big now)

Changed layout to be called frontend just because render is already an identifier.

Loads statuspage-widget but as a dynamic import, so it shouldn't hold up the page load.