OE-135 Correct and predictable report test.
Right, this has been boggling my mind a bit as I understand what's going on.

Firstly the window being 24 hours has increased the expected active things "today". Secondly there is some flappiness which turns out to be because one of the assessments has a 90 minute adjustment and a start time that takes its finish time to exactly the mock current time of 10:30am, and timing decides whether that is just before or after now.

The other flappiness was because the makeStoredStudentAssessment fixture had a nondeterministic random call in it, meaning you'd get different adjustments every time you ran, which was awesome. Changed to use the student ID as a seed, so although random it's at least the same random every time (def randomDiceRoll = 4)