OE-285 Handle start time being slightly in the future.
Wondering if it's ever a good idea to render this countdown based on the browser's clock, or if we should always wait until we've received a timestamp from the websocket. Alternatively we should include the server's timestamp in the data attributes and use that to form a time delta whenever we're rendering based on the user's clock, so even if it's fast or slow it would still be able to render accurately offline.