OE-327 Send all announcements on heartbeat.
Which includes on connect since we do an initial heartbeat.

Body of the exam page now has a data-assessment-id attribute which could be hooked into by the websocket to fix the problem where we're subscribing to too many topics. At the moment there is just a quick fix to filter down to in-progress assessments, but that still means if you've started two assessments you can be seeing announcements for one on the other's page. Will Jira that separately.