OE-207 Upgrade play-utils.
When I had added local changes to make the affected tests fail every time (which I'm not committing because it involved a System.gc inside the service method), this upgrade fixed them. This should mean that it will fix the occasional failures in AdminAssessmentsControllerSpec and the AssessmentSpec integration test too.